Разбил навигационное меню на компоненты
